eclipse suggested parenthesis
The eclipse IDE is a wonderfull tool to research and investigate foreign code, although it is not quite optimal for normal development. The codeanalysis feature of eclipse spits out a list of potential problems and make suggestions. One of the suggestions that eclipse made for libapenetwork was that parentheses should be used if bitwise operations take place.