solved some compile warnings for -Wmissing-declarations
One function was used internally but not declared static.
One function was not defined in the header file although it was a public (APE_...)
One function was static but not used, because it was in a comment block.
Currently there is one more warning, but that will be solved by a8c152f4d59980ba2a738242719776520449414d.